Merge with live version and begin translation
The site now corresponds to the live version at landscapelab.boku.ac.at again. Some initial work has been done on translating as a proof-of-concept.

Some files were moved around in the process, mainly because I had to manually merge the live version with this repository. Everything is working now, but it's possible that a directory got an unintended structure or that files were renamed needlessly.